+++ src/zope/publisher/publish.py
@@ -25,6 +25,7 @@ from zope.proxy import removeAllProxies
 
 _marker = object()  # Create a new marker object.
 
+
 def unwrapMethod(obj):
     """obj -> (unwrapped, wrapperCount)
 
@@ -41,14 +42,14 @@ def unwrapMethod(obj):
         if bases is not None:
             raise TypeError("mapply() can not call class constructors")
 
-        im_func = getattr(unwrapped, '__func__', None)
+        im_func = getattr(unwrapped, six._meth_func, None)
         if im_func is not None:
             unwrapped = im_func
             wrapperCount += 1
-        elif getattr(unwrapped, '__code__', None) is not None:
+        elif getattr(unwrapped, six._func_code, None) is not None:
             break
         else:
-            unwrapped = getattr(unwrapped, '__call__' , None)
+            unwrapped = getattr(unwrapped, '__call__', None)
             if unwrapped is None:
                 raise TypeError("mapply() can not call %s" % repr(obj))
     else: