+++ src/zope/publisher/publish.py
@@ -42,11 +42,16 @@ def unwrapMethod(obj):
         if bases is not None:
             raise TypeError("mapply() can not call class constructors")
 
-        im_func = getattr(unwrapped, six._meth_func, None)
+        im_func = getattr(unwrapped, '__func__', None)
+        if im_func is None:
+            # Backwards compatibility with objects aimed at Python 2
+            im_func = getattr(unwrapped, 'im_func', None)
         if im_func is not None:
             unwrapped = im_func
             wrapperCount += 1
-        elif getattr(unwrapped, six._func_code, None) is not None:
+        elif getattr(unwrapped, '__code__', None) is not None:
+            break
+        elif getattr(unwrapped, 'func_code', None) is not None:
             break
         else:
             unwrapped = getattr(unwrapped, '__call__', None)