+++ CHANGES.rst
@@ -4,7 +4,14 @@ Changes
 4.3.1 (unreleased)
 ------------------
 
-- TBD
+- Fix file uploads on python 3.4 and up. cgi.FieldStorage explicitly
+  closes files when it is garbage collected. For details, see:
+
+  * http://bugs.python.org/issue18394
+  * https://hg.python.org/cpython/rev/c0e9ba7b26d5
+
+  We now keep a reference to the FieldStorage till we are finished
+  processing the request.
 
 
 4.3.0 (2016-07-04)